+++ lib/ui/compositing/scene_builder.h
@@ -46,7 +46,7 @@ class SceneBuilder : public ftl::RefCountedThreadSafe<SceneBuilder>,
                       double maskRectTop,
                       double maskRectBottom,
                       int blendMode);
-  void pushPhysicalModel(const RRect& rrect, int elevation, int color);
+  void pushPhysicalModel(const RRect& rrect, double elevation, int color);
 
   void pop();
